Move fastapi adapter according to hexkit architecture (#10)
Move fastapi adapter according to hexkit architecture
