+++ tests/integration/test_post_endpoint.py
@@ -20,8 +20,8 @@ import crypt4gh.header
 import pytest
 from fastapi.testclient import TestClient
 
-from ekss.api.deps import config_injector
-from ekss.api.main import setup_app
+from ekss.adapters.inbound.fastapi_.deps import config_injector
+from ekss.adapters.inbound.fastapi_.main import setup_app
 from ekss.config import CONFIG
 from tests.fixtures.file import first_part_fixture  # noqa: F401
 from tests.fixtures.file import FirstPartFixture
